requirements.txt
colors
ansicolors==1.1.8
apex==0.1
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.32
pandas==1.0.1
paramiko==2.7.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-fix==2.0.0.1
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
storage==0.0.4.3
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.43.0
uritemplate==3.0.1
urllib3==1.25.8
